combine the books and categories reducers
