+++ src/redux/books/books.js
@@ -1,11 +1,28 @@
 import ActionType from '../categorie/categories';
 
-export const addBook = (books) => ({
+const initialState = [];
+
+export const addBook = (payload) => ({
   type: ActionType.ADD_BOOK,
-  payload: books,
+  payload,
 });
 
-export const removeBook = (books) => ({
+export const removeBook = (payload) => ({
   type: ActionType.REMOVE_BOOK,
-  payload: books,
+  payload,
 });
+
+const booksReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.ADD_BOOK:
+      return [...state, action.payload];
+
+    case ActionType.REMOVE_BOOK:
+      return state.filter((book) => book.id !== action.payload);
+
+    default:
+      return state;
+  }
+};
+
+export default booksReducer;
